0.2.0 audit fixes: README refresh, no em-dashes, float popup seconds, deck data dir
Address the pre-tag audit: bring README up to date with the Stream Deck,
Home Assistant camera pop-ups, fullscreen audio, glance.local naming, and
touchscreen calibration; remove em-dashes from the Stream Deck README
(project style); accept a fractional popup duration on
/events/camera-popup so a Stream Deck press is never rejected; and have
the installer create and own the deck's data dir so it can record its
selection.

+++ service/app/routers/events.py
@@ -40,7 +40,9 @@ _DEFAULT_POPUP_SECONDS = 20
 
 class CameraPopupPayload(BaseModel):
     camera: str = ""             # camera id or name; empty = the first camera
-    seconds: int = 0             # 0 = the default popup duration
+    # Accept a float so a caller (e.g. the Stream Deck, whose duration is a
+    # float) is never rejected; clamp_seconds rounds it to whole seconds.
+    seconds: float = 0           # 0 = the default popup duration
 
 
 class NotifyPayload(BaseModel):
